Rename currentUser.ts store to session.ts
