+++ src/store/index.ts
@@ -1,4 +1,4 @@
-import currentUserStore from './currentUser';
+import currentUserStore from './session';
 import documentsStore from './documents';
 import attributesStore from './attributes';
 import docTypesStore from './docTypes';
